09-AI_Agents/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
chromadb
unstructured
tiktoken
python-dotenv
langchain-google-genai
sentence-transformers
